chore(shared): merge some modules
