--- packages/shared/src/apply/hmr.ts
@@ -1,14 +1,0 @@
-import { ModifyChainUtils, NormalizedConfig } from '../types';
-
-export function isUsingHMR(
-  config: NormalizedConfig,
-  { isProd, target }: Pick<ModifyChainUtils, 'isProd' | 'target'>,
-) {
-  return (
-    !isProd &&
-    target !== 'node' &&
-    target !== 'web-worker' &&
-    target !== 'service-worker' &&
-    config.dev.hmr
-  );
-}
