--- packages/shared/src/apply/index.ts
@@ -1,5 +1,0 @@
-export * from './html';
-export * from './output';
-export * from './resolve';
-export * from './hmr';
-export * from './babel';
